orden de carpetas, restructuracion de componentes
