+++ src/css/Login.css
@@ -1,91 +1,131 @@
-.container{
-    position: relative;
-    left: 40px;
-    top: -2px;
-    width: 300%;
-    height: 200%;
-    color: gray;
-
+.container {
+  position: relative;
+  left: 40px;
+  top: -2px;
+  width: 300%;
+  height: 200%;
+  color: gray;
 }
 
-.container-fluid{
-    color: white ;
+.container-fluid {
+  color: white;
 }
 
-.letras{
-    color: white;
-    font-weight: bold;
-    margin-bottom: 0;
-    font-size:25px ;
+.letras {
+  color: white;
+  font-weight: bold;
+  margin-bottom: 0;
+  font-size: 25px;
 }
 
-.letras-r{
-    color: #3F69C8;
-    font-weight: bold;
-    margin-bottom: 0;
-    font-size:35px ;
+.letras-r {
+  color: #3f69c8;
+  font-weight: bold;
+  margin-bottom: 0;
+  font-size: 35px;
 }
 
-.img-fluid{
-    position: relative;
-    left: 40px;
-    top: -2px;
-    width: 300%;
-    height: 200%;
-    color: gray;
+.img-fluid {
+  position: relative;
+  left: 40px;
+  top: -2px;
+  width: 300%;
+  height: 200%;
+  color: gray;
 }
-.form-label{
-    color: #3F69C8;
-    font-weight: bold;
+.form-label {
+  color: #3f69c8;
+  font-weight: bold;
 }
-.bs-card-bg{
-    color:#D9D9D9;
+.bs-card-bg {
+  color: #d9d9d9;
 }
 .custom-nav {
-    background-color: #D9D9D9;
-  }
+  background-color: #d9d9d9;
+}
 
-  .producto {
-    width: 100%;
-    height: auto;
-  }
-  /* Usuario.css */
+.producto {
+  width: 100%;
+  height: auto;
+}
+/* Usuario.css */
 
 .usuario {
-    margin-top: 20px;
-  }
-  
-  .catalogo {
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    gap: 20px;
-  }
-  
-  .producto {
-    text-align: center;
-  }
-  
-  .producto img {
-    width: 50%;
-    height: auto;
-  }
-  
-  .producto h3 {
-    margin-top: 10px;
-  }
-  
-  .producto p {
-    margin-bottom: 10px;
-  }
-  
-  .producto button {
-    background-color: #3f69c8;
-    color: white;
-    padding: 5px 10px;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-  }
-  
+  margin-top: 20px;
+}
 
+.catalogo {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 20px;
+}
 
+.producto {
+  text-align: center;
+}
+
+.producto img {
+  width: 50%;
+  height: auto;
+}
+
+.producto h3 {
+  margin-top: 10px;
+}
+
+.producto p {
+  margin-bottom: 10px;
+}
+
+.producto button {
+  background-color: #3f69c8;
+  color: white;
+  padding: 5px 10px;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+}
+.footer {
+  background-color: #3f69c8;
+  position: relative;
+  padding: 20px 0;
+  margin-top: 50px;
+}
+
+.footer h3 {
+  color: white;
+}
+
+.footer-form {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  margin-top: 20px;
+}
+
+.input-group {
+  display: flex;
+  justify-content: flex-start;
+  align-items: flex-start;
+}
+
+.form-contro {
+ 
+  width: 246px;
+  height: 39px;
+  left: 94px;
+  
+  background: #C6C6C6;
+  border: 2px solid #FFFFFF;
+  border-radius: 6px;
+}
+
+.btn-primary {
+  
+  width: 69px;
+  left: 350px;
+  height: 39px;
+  background: #c6c6c6;
+  border: 2px solid #ffffff;
+  border-radius: 6px;
+}
